[NFC] Move StorageRefResult to a header
