+++ include/swift/AST/DiagnosticsSIL.def
@@ -1225,6 +1225,6 @@ NOTE(lifetime_outside_scope_escape, none,
 
 ERROR(noncopyable_shared_case_block_unimplemented, none,
       "matching a non-'Copyable' value using a case label that has multiple patterns is not implemented", ())
-
+      
 #define UNDEFINE_DIAGNOSTIC_MACROS
 #include "DefineDiagnosticMacros.h"
